requirements.txt
google-generativeai==0.3.2
PyPDF2==3.0.1
networkx==3.1
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.24.3